requirements.txt
numpy
scipy
Pillow
sklearn
matplotlib
opencv-python